// src/frontend/src/backend.d.ts
export type Time = bigint;
export interface Feedback {
    id: bigint;
    status: FeedbackStatus;
    title: string;
    createdAt: Time;
    authorName?: string;
    message: string;
    category: FeedbackCategory;
    authorPrincipal?: Principal;
}
export interface Language {
    flagEmoji: string;
    code: string;
    name: string;
    createdAt: Time;
    gradientStart: string;
    ordering: bigint;
    textDirection: TextDirection;
    gradientEnd: string;
}
export interface Word {
    id: bigint;
    difficulty: Difficulty;
    languageName: string;
    addedAt: Time;
    foreign: string;
    examples: Array<string>;
    english: string;
}
// ...
export enum Difficulty {
    beginner = "beginner",
    advanced = "advanced",
    hard = "hard",
    medium = "medium"
}
export enum DifficultySelector {
    all = "all",
    beginner = "beginner",
    advanced = "advanced",
    hard = "hard",
    medium = "medium"
}
export enum FeedbackCategory {
    bug = "bug",
    idea = "idea",
    issue = "issue"
}
export enum FeedbackStatus {
    pending = "pending",
    completed = "completed"
}
export enum TextDirection {
    ltr = "ltr",
    rtl = "rtl"
}